requirements.txt
elasticsearch>=8.0.0
python-dotenv
numpy<2.0
torch>=2.4.0
sentence-transformers>=3.0.0
transformers>=4.41.0,<5.0.0
pymupdf
pandas
openpyxl
tqdm
requests
streamlit
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
mcp[cli]>=1.26.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
aiohttp>=3.9.0
